requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
pydantic
diskcache
prometheus-client
pillow
